+++ decoder/jpeg2000/jp2_nocgo.go
@@ -15,7 +15,7 @@ func init() {
 type stubFactory struct{}
 
 func (f *stubFactory) Name() string                  { return "jpeg2000" }
-func (f *stubFactory) TIFFCompressionTags() []uint16 { return []uint16{33003, 34712} }
+func (f *stubFactory) TIFFCompressionTags() []uint16 { return []uint16{33003, 33005, 34712} }
 func (f *stubFactory) New() decoder.Decoder          { return &stubDecoder{} }
 
 type stubDecoder struct{}